Added small delay right after adding data to ES to allow time for the index to be populated
